A bit more security refactor
