--- src/main/java/es/urjc/etsii/grafo/iudex/security/ConfigSecurity.java
@@ -1,34 +1,0 @@
-package es.urjc.etsii.grafo.iudex.security;
-
-import es.urjc.etsii.grafo.iudex.security.jwt.JwtRequestFilter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity(jsr250Enabled=true)
-class ConfigSecurity {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http, JwtRequestFilter jwtRequestFilter) throws Exception {
-        http.authorizeHttpRequests(authorize -> authorize.anyRequest().permitAll());
-
-        http.oauth2Login(oauth2 -> oauth2.defaultSuccessUrl("/API/v1/oauth/login", true));
-
-        // Disable CORS and CSRF protection
-        http.cors(AbstractHttpConfigurer::disable).csrf(AbstractHttpConfigurer::disable);
-
-        // Add JWT Token filter
-        http.addFilterBefore(jwtRequestFilter, BasicAuthenticationFilter.class);
-
-        return http.build();
-    }
-
-}
