+++ src/main/java/es/urjc/etsii/grafo/iudex/services/UserService.java
@@ -3,9 +3,9 @@ package es.urjc.etsii.grafo.iudex.services;
 import es.urjc.etsii.grafo.iudex.entities.User;
 import es.urjc.etsii.grafo.iudex.exceptions.IudexException;
 import es.urjc.etsii.grafo.iudex.repositories.UserRepository;
-import es.urjc.etsii.grafo.iudex.security.jwt.AuthResponse;
-import es.urjc.etsii.grafo.iudex.security.jwt.JwtTokenProvider;
-import es.urjc.etsii.grafo.iudex.security.jwt.UserDetailsServiceImp;
+import es.urjc.etsii.grafo.iudex.security.AuthResponse;
+import es.urjc.etsii.grafo.iudex.security.JwtTokenProvider;
+import es.urjc.etsii.grafo.iudex.security.UserDetailsServiceImp;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
