+++ src/test/java/es/urjc/etsii/grafo/iudex/integrationtest/api_controllers/TestAPIAdminController.java
@@ -3,11 +3,10 @@ package es.urjc.etsii.grafo.iudex.integrationtest.api_controllers;
 import es.urjc.etsii.grafo.iudex.api.v1.APIAdminController;
 import es.urjc.etsii.grafo.iudex.entities.Result;
 import es.urjc.etsii.grafo.iudex.entities.Sample;
-import es.urjc.etsii.grafo.iudex.security.jwt.JwtRequestFilter;
+import es.urjc.etsii.grafo.iudex.security.JwtRequestFilter;
 import es.urjc.etsii.grafo.iudex.services.ResultService;
 import es.urjc.etsii.grafo.iudex.utils.JSONConverter;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
